Singularity based RStudio
This commit removes the requirement for using proot to create private /tmp spaces. Instead a Singularity container is used which mounts RStudio and its dependencies from the host, plus overrides the mount for /tmp using a new tmp dir rooted in PBSTMP. The Singularity container is as simple as can be requiring only the installation of `which` over the base CentOS 7.

The path to the image is currently in my home directory, and needs to be moved to a system directory, or the WIAG user's home.