Implementation of a hybrid vnc / basic RStudio
A user requested the ability to be able to open X11 windows from their RStudio session. In order to support this request it is necessary to switch from RStudio Server to the desktop version of the application, running within a full desktop environment. This commit offers the user the ability at batch-submit-time to select whether they want RStudio Desktop or Server. Pleasingly RStudio Desktop respects the TMPDIR setting, and does not need to be contained within a Singularity 'chroot'. Unfortunately because of the copy and paste limitations of our VNC implementation there is a significant difference in the user experience between the two modes, preventing RStudio desktop from being a clear favorite.

After consultation with @summerwang it was decided that this implementation is not required, as users may alternatively work from a regular BatchConnect desktop session.

In order to offer a BatchConnect application that can run in 'basic' or 'vnc' modes, where the 'basic' mode requires a custom launch button (view.html.erb), it was necessary to change the batchconnect/session/panel view in the Dashboard. If support for dual mode BatchConnect applications is desired in the future it would be advisable to move the template selection logic out of the template.