use centos7 img and force home=sing_home
there is some issue in binding all of etc /etc (which this image does)
and so we dropped it. only we need all of /etc so this re-enables it.
the isse is fixed if we set HOME=SINGULARITY_HOME so users in tutorials
and classes can create a ~/.rstudio directory (which before wouldn't work
because HOME was still being evaluated as the actual home which didn't
exist in the container).